// src/app/projects/project.interface.ts
export interface Project {
    title: string;
    image: string;
    galleryTitle: string[];
    gallery: string[];
    technologies: string[];
    galeryId: string;
    github: string;
    externalLink?: string;
    description: string;
    idLink: string;
    idImage: string;
    descriptionIsShow: boolean;
  }